backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-pptx==1.0.2
psycopg2-binary==2.9.10
openai==1.59.3
python-dotenv==1.0.1
werkzeug==3.1.3
